pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evaldeck"
version = "0.1.0"
description = "The evaluation framework for AI agents. Pytest for agents."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "Evaldeck Contributors" }]
keywords = [
    "ai",
    "agents",
    "evaluation",
    "testing",
    "llm",
    "langchain",
    "crewai",
    "autogen",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "opentelemetry-sdk>=1.20",
    "openinference-semantic-conventions>=0.1",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.18"]
all = ["evaldeck[openai,anthropic]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
    "mkdocs-autorefs>=0.5",
]

[project.scripts]
evaldeck = "evaldeck.cli:main"

[project.urls]
Homepage = "https://github.com/evaldeck-ai/evaldeck"
Documentation = "https://evaldeck.dev"
Repository = "https://github.com/evaldeck-ai/evaldeck"
Issues = "https://github.com/evaldeck-ai/evaldeck/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/evaldeck"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
